pyproject.toml
[build-system]
requires=["setuptools","setuptools-scm"]
build-backend= "setuptools.build_meta"

[project]
name="aivr"
version="0.0.10"
dependencies=[ "numpy==1.26.0", "opencv-python==4.5.5.62", "pyzmq==25.1.1","sys","json","time", "os","threading"]
license = {file = "LICENSE"}
authors = [
    { name = "Waqas Kureshy", email = "waqas.kureshy319@gmail.com" },
]
